refactor: rename yubikey2 as yubikey
